requirements.txt
fastapi==0.61.1
uvicorn==0.12.1
python-multipart==0.0.5
python-jose[cryptography]
passlib[bcrypt]
email-validator==1.1.1
pytest==6.1.1
requests==2.24.0